+++ pystarport/cosmoscli.py
@@ -68,6 +68,16 @@ class ChainCommand:
 
         return "Available Commands" in output.decode()
 
+    def prob_tendermint_subcommand(self):
+        'test if the command has "tendermint" subcommand, removed in evm 0.4.x'
+        try:
+            output = self("tendermint")
+        except AssertionError:
+            # non-zero return code
+            return False
+
+        return "Available Commands" in output.decode()
+
     def __call__(self, cmd, *args, stdin=None, stderr=subprocess.STDOUT, **kwargs):
         "execute chain-maind"
         args = " ".join(build_cli_args_safe(cmd, *args, **kwargs))
@@ -100,10 +110,12 @@ class CosmosCLI:
         self.error = None
         self.has_genesis_subcommand = self.raw.prob_genesis_subcommand()
         self.has_icaauth_subcommand = self.raw.prob_icaauth_subcommand()
+        self.has_tendermint_subcommand = self.raw.prob_tendermint_subcommand()
 
     def node_id(self):
         "get tendermint node id"
-        output = self.raw("tendermint", "show-node-id", home=self.data_dir)
+        subcmd = "tendermint" if self.has_tendermint_subcommand else "comet"
+        output = self.raw(subcmd, "show-node-id", home=self.data_dir)
         return output.decode().strip()
 
     def delete_account(self, name):
@@ -119,8 +131,10 @@ class CosmosCLI:
             keyring_backend="test",
         )
 
-    def create_account(self, name, mnemonic=None):
+    def create_account(self, name, mnemonic=None, **kwargs):
         "create new keypair in node's keyring"
+        if kwargs.get("coin_type") == 60:
+            kwargs["key_type"] = "eth_secp256k1"
         if mnemonic is None:
             output = self.raw(
                 "keys",
@@ -129,6 +143,7 @@ class CosmosCLI:
                 home=self.data_dir,
                 output="json",
                 keyring_backend="test",
+                **kwargs,
             )
         else:
             output = self.raw(
@@ -140,11 +155,14 @@ class CosmosCLI:
                 output="json",
                 keyring_backend="test",
                 stdin=mnemonic.encode() + b"\n",
+                **kwargs,
             )
         return json.loads(output)
 
-    def create_account_ledger(self, name):
+    def create_account_ledger(self, name, **kwargs):
         "create new ledger keypair"
+        if kwargs.get("coin_type") == 60:
+            kwargs["key_type"] = "eth_secp256k1"
 
         def send_request():
             try:
@@ -156,6 +174,7 @@ class CosmosCLI:
                     home=self.data_dir,
                     output="json",
                     keyring_backend="test",
+                    **kwargs,
                 )
             except Exception as e:
                 self.error = e
@@ -277,19 +296,32 @@ class CosmosCLI:
                 node=self.node_rpc,
             )
         )["commission"]
-        return parse_amount((res.get("commission") or res)[0])
+        if isinstance(res, dict):
+            res = res["commission"]
+        return parse_amount(res[0])
 
-    def distribution_community(self):
-        res = json.loads(
-            self.raw(
-                "query",
-                "distribution",
-                "community-pool",
-                output="json",
-                node=self.node_rpc,
-            )
-        )
-        return parse_amount(res["pool"][0])
+    def distribution_community(self, **kwargs):
+        for module in ["distribution", "protocolpool"]:
+            try:
+                res = json.loads(
+                    self.raw(
+                        "query",
+                        module,
+                        "community-pool",
+                        output="json",
+                        node=self.node_rpc,
+                        **kwargs,
+                    )
+                )
+                return parse_amount(res["pool"][0])
+            except Exception as e:
+                if (
+                    module == "distribution"
+                    and "CommunityPool query exposed by the external community pool"
+                    in str(e)
+                ):
+                    continue
+                raise
 
     def distribution_reward(self, delegator_addr):
         res = json.loads(
@@ -386,7 +418,11 @@ class CosmosCLI:
                 **kwargs,
             )
         )
-        if not generate_only and rsp["code"] == 0 and event_query_tx:
+        if (
+            not generate_only
+            and rsp["code"] == 0
+            and event_query_tx
+        ):
             rsp = self.event_query_tx_for(rsp["txhash"])
         return rsp
 
@@ -422,7 +458,11 @@ class CosmosCLI:
                         **kwargs,
                     )
                 )
-                if not generate_only and self.output["code"] == 0 and event_query_tx:
+                if (
+                    not generate_only
+                    and self.output["code"] == 0
+                    and event_query_tx
+                ):
                     self.output = self.event_query_tx_for(self.output["txhash"])
             except Exception as e:
                 self.error = e
@@ -555,7 +595,7 @@ class CosmosCLI:
             rsp = self.event_query_tx_for(rsp["txhash"])
         return rsp
 
-    def make_multisig(self, name, signer1, signer2):
+    def make_multisig(self, name, signer1, signer2, **kwargs):
         self.raw(
             "keys",
             "add",
@@ -564,6 +604,7 @@ class CosmosCLI:
             multisig_threshold="2",
             home=self.data_dir,
             keyring_backend="test",
+            **kwargs,
         )
 
     def sign_multisig_tx(self, tx_file, multi_addr, signer_name, **kwargs):
@@ -714,9 +755,10 @@ class CosmosCLI:
         } | options
 
         if "pubkey" not in options:
+            subcmd = "tendermint" if self.has_tendermint_subcommand else "comet"
             pubkey = (
                 self.raw(
-                    "tendermint",
+                    subcmd,
                     "show-validator",
                     home=self.data_dir,
                 )
@@ -760,9 +802,10 @@ class CosmosCLI:
     ):
         """MsgCreateValidator
         create the node with create_node before call this"""
+        subcmd = "tendermint" if self.has_tendermint_subcommand else "comet"
         pubkey = (
             self.raw(
-                "tendermint",
+                subcmd,
                 "show-validator",
                 home=self.data_dir,
             )
@@ -1445,3 +1488,16 @@ class CosmosCLI:
                 output="json",
             )
         )["denom_trace"]
+
+    def ibc_denom(self, path, node):
+        denom_hash = hashlib.sha256(path.encode()).hexdigest().upper()
+        return json.loads(
+            self.raw(
+                "q",
+                "ibc-transfer",
+                "denom",
+                denom_hash,
+                node=node,
+                output="json",
+            )
+        )["denom"]